+++ sw.js
@@ -1,6 +1,6 @@
 /*global self, caches*/
 /*jshint esversion: 6*/
-var snapVersion = '11.0.5',
+var snapVersion = '11.0.6',
     cacheName = `snap-pwa-${snapVersion}`,
     filesToCache = [
         'snap.html',