+++ HISTORY.md
@@ -5,7 +5,8 @@
     * fixed default values for color slots & improved default slot values overall
 
 ### 2025-10-21
-    * objects, blocks: fixed default values for color slots & improved default slot values overall
+* objects, blocks: fixed default values for color slots & improved default slot values overall
+* objects: (internal) methods to externally persist and restore which global blocks (primitives, custom blocks, variables) are visible in the palette
 
 ## 11.0.5:
 * **Notable Changes:**